[SCB-2855]make ExceptionToProducerResponseConverters extensible
